Support deploying when multiple backends are used
If no backend besides the HTML one is configured, mdbook builds the book
inside `book/`. If another backend is configured, even if it doesn't
run, the book is built inside `book/html/`. Given that changes in the
source can lead to the removal of some backend at some point in time,
this makes the deployment resistant to those changes.

Deleting the `book/` dir helps in avoiding leftover directories or files
affecting the deployment process.